+++ integration/failure/module-init-error/test/src/ModuleInitErrorTests.scala
@@ -24,6 +24,22 @@ object ModuleInitErrorTests extends IntegrationTestSuite {
         res2.out.contains("foo.fooTarget"),
         res2.out.contains("foo.fooCommand")
       )
+
+      val res3 = evalStdout("resolve", "__.fooTarget")
+      assert(res3.isSuccess == true)
+      assert(res3.out.contains("foo.fooTarget"))
+
+      val res4 = evalStdout("resolve", "__")
+      assert(res4.isSuccess == true)
+      assert(res4.out.contains("bar"))
+      assert(res4.out.contains("bar.barCommand"))
+      assert(res4.out.contains("bar.barTarget"))
+      assert(res4.out.contains("bar.qux"))
+      assert(res4.out.contains("bar.qux.quxCommand"))
+      assert(res4.out.contains("bar.qux.quxTarget"))
+      assert(res4.out.contains("foo"))
+      assert(res4.out.contains("foo.fooCommand"))
+      assert(res4.out.contains("foo.fooTarget"))
     }
 
     test("rootTarget") {
