+++ main/define/src/mill/define/Module.scala
@@ -2,11 +2,8 @@ package mill.define
 
 import mill.api.internal
 
-import java.lang.reflect.Modifier
-
 import scala.language.experimental.macros
 import scala.reflect.ClassTag
-import scala.reflect.NameTransformer.decode
 
 /**
  * Represents a namespace within the Mill build hierarchy, containing nested
@@ -45,80 +42,6 @@ class Module(implicit outerCtx0: mill.define.Ctx) extends mill.moduledefs.Cacher
 }
 
 object Module {
-  @internal
-  object Internal {
-    import fastparse._, NoWhitespace._
-
-    def ident[_p: P]: P[String] = P(CharsWhileIn("a-zA-Z0-9_\\-")).!
-
-    def standaloneIdent[_p: P]: P[String] = P(Start ~ ident ~ End)
-
-    def isLegalIdentifier(identifier: String): Boolean =
-      parse(identifier, standaloneIdent(_)).isInstanceOf[Parsed.Success[_]]
-
-    def reflect(
-        outer: Class[_],
-        inner: Class[_],
-        filter: String => Boolean,
-        noParams: Boolean
-    ): Seq[java.lang.reflect.Method] = {
-      for {
-        m <- outer.getMethods.sortBy(_.getName)
-        n = decode(m.getName)
-        if filter(n) &&
-          isLegalIdentifier(n) &&
-          (!noParams || m.getParameterCount == 0) &&
-          (m.getModifiers & Modifier.STATIC) == 0 &&
-          (m.getModifiers & Modifier.ABSTRACT) == 0 &&
-          inner.isAssignableFrom(m.getReturnType)
-      } yield m
-    }
-
-    // For some reason, this fails to pick up concrete `object`s nested directly within
-    // another top-level concrete `object`. This is fine for now, since Mill's Ammonite
-    // script/REPL runner always wraps user code in a wrapper object/trait
-    def reflectNestedObjects[T: ClassTag](
-        outer: Class[_],
-        filter: String => Boolean = Function.const(true)
-    ): Seq[java.lang.reflect.Member] = {
-      reflect(outer, classOf[Object], filter, noParams = true) ++
-        outer
-          .getClasses
-          .filter(implicitly[ClassTag[T]].runtimeClass.isAssignableFrom(_))
-          .flatMap(c =>
-            c.getFields.find(_.getName == "MODULE$")
-          ).distinct
-    }
-
-    def reflectNestedObjects0[T: ClassTag](outerCls: Class[_], filter: String => Boolean = Function.const(true))
-    : Seq[(String, java.lang.reflect.Member)] = {
-
-      val first = Module.Internal
-        .reflect(
-          outerCls,
-          implicitly[ClassTag[T]].runtimeClass,
-          filter,
-          noParams = true
-        )
-        .map(m => (m.getName, m))
-
-      val second =
-        outerCls
-          .getClasses
-          .filter(implicitly[ClassTag[T]].runtimeClass.isAssignableFrom(_))
-          .flatMap { c =>
-            c.getName.stripPrefix(outerCls.getName) match {
-              case s"$name$$" if filter(name) =>
-                c.getFields.find(_.getName == "MODULE$").map(name -> _)
-              case _ => None
-            }
-
-          }
-          .distinct
-
-      first ++ second
-    }
-  }
 
   @internal
   class Internal(outer: Module) {
@@ -133,18 +56,8 @@ object Module {
     lazy val targets: Set[Target[_]] =
       traverse { _.millInternal.reflectAll[Target[_]].toIndexedSeq }.toSet
 
-    lazy val segmentsToTargets: Map[Segments, Target[_]] = targets
-      .map(t => (t.ctx.segments, t))
-      .toMap
-
-    // Ensure we do not propagate the implicit parameters as implicits within
-    // the body of any inheriting class/trait/objects, as it would screw up any
-    // one else trying to use sourcecode.{Enclosing,Line} to capture debug info
-    lazy val millModuleEnclosing: String = outer.millOuterCtx.enclosing
-    lazy val millModuleLine: Int = outer.millOuterCtx.lineNum
-
     def reflect[T: ClassTag](filter: String => Boolean): Seq[T] = {
-      Module.Internal.reflect(
+      Reflect.reflect(
         outer.getClass,
         implicitly[ClassTag[T]].runtimeClass,
         filter,
@@ -155,8 +68,6 @@ object Module {
 
     def reflectAll[T: ClassTag]: Seq[T] = reflect[T](Function.const(true))
 
-    def reflectSingle[T: ClassTag](label: String): Option[T] = reflect(_ == label).headOption
-
     // For some reason, this fails to pick up concrete `object`s nested directly within
     // another top-level concrete `object`. This is fine for now, since Mill's Ammonite
     // script/REPL runner always wraps user code in a wrapper object/trait
@@ -166,7 +77,7 @@ object Module {
 
     def reflectNestedObjects0[T: ClassTag](filter: String => Boolean = Function.const(true))
         : Seq[(String, () => T)] = {
-      Internal.reflectNestedObjects0(outer.getClass).map{
+      Reflect.reflectNestedObjects0(outer.getClass, filter).map{
         case (name, m: java.lang.reflect.Method) => (name, () => m.invoke(outer).asInstanceOf[T])
         case (name, m: java.lang.reflect.Field) => (name, () =>m.get(outer).asInstanceOf[T])
       }
